feat(terminal): better controls for login
